requirements.txt
oracledb
python-dotenv
requests
